+++ src/Filter/Modifier/PriceModifierInterface.php
@@ -8,6 +8,6 @@ use App\Entity\Promotion;
 interface PriceModifierInterface
 {
 
-    public function modify(float $price, int $quantity, Promotion $promotion, PromotionEnquiryInterface $enquiry): float;
+    public function modify(int $price, int $quantity, Promotion $promotion, PromotionEnquiryInterface $enquiry): int;
 
 }
